requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
kafka-python
apscheduler
alembic
pytest
